fix to float

+++ app/api_v1/nodes/schemas.py
@@ -2,9 +2,9 @@ from pydantic import BaseModel, ConfigDict
 from datetime import datetime
 
 class Node(BaseModel):
-    x: int
-    y: int
-    z: int
+    x: float
+    y: float
+    z: float
     connections: list[str]
     floor: int
     pid_name: str
@@ -27,9 +27,9 @@ class NodeFilter(BaseModel):
     id: int | None = None
     created_at: datetime | None = None
     updated_at: datetime | None = None
-    x_coord: int | None = None
-    y_coord: int | None = None
-    z_coord: int | None = None
+    x_coord: float | None = None
+    y_coord: float | None = None
+    z_coord: float | None = None
     connections: list[str] | None = None
     floor: int | None = None
     pid_name: str | None = None
